[*] Characters requests moved to store
