+++ src/pages/characters/Characters.tsx
@@ -2,8 +2,6 @@ import {FC, useEffect} from "react";
 import styles from "./styles.module.css"
 import Card from "../../components/card/Card.tsx";
 import SearchBar from "../../components/searchBar/SearchBar.tsx";
-import charactersRequests from "../../api/charactersRequests.ts"
-import {ICharacter} from "../../types/ICharacter.tsx";
 import Pagination from "../../components/pagination/Pagination.tsx";
 import useDebounce from "../../hooks/useDebounce.tsx";
 import {ColorRing} from "react-loader-spinner";
@@ -12,27 +10,13 @@ import {observer} from "mobx-react-lite";
 
 const Characters: FC = observer(() => {
     const debouncedInput = useDebounce(store.searchCharacter, 3000);
-    const charactersOnPage : number = 36;
 
     useEffect(() => {
-        charactersRequests.getCharacters(charactersOnPage,(store.page - 1) * charactersOnPage)
-            .then(data => {
-                const charactersArray: ICharacter[]  = data.data.results.map(character => {
-                    return {
-                        id: character.id,
-                        name: character.name,
-                        desc: character.description,
-                        image: character.thumbnail.path + "." + character.thumbnail.extension
-                    }
-                })
-                store.setCharacters(charactersArray);
-                store.setCharactersAmount(data.data.total);
-                store.setPagesAmount(Math.ceil(store.charactersAmount / charactersOnPage));
-            })
+        store.searchCharacters((store.page - 1) * store.charactersOnPage);
     }, [store.page]);
 
     useEffect(() => {
-        if (debouncedInput) searchCharactersByNameDebounce();
+        if (debouncedInput) store.searchCharactersByNameDebounce();
         console.log("debounce");
     }, [debouncedInput]);
 
@@ -44,40 +28,9 @@ const Characters: FC = observer(() => {
         store.setSearchCharacter(event.target.value);
     }
 
-    const searchCharactersByName = (event: any) => {
-        event.preventDefault();
-        charactersRequests.searchCharacterByName(charactersOnPage, store.searchCharacter)
-        .then(data => {
-            const charactersArray: ICharacter[]  = data.data.results.map(character => {
-                return {
-                    id: character.id,
-                    name: character.name,
-                    desc: character.description,
-                    image: character.thumbnail.path + "." + character.thumbnail.extension
-                }
-            })
-            store.setCharacters(charactersArray);
-        })
-    }
-
-    const searchCharactersByNameDebounce = () => {
-        charactersRequests.searchCharacterByName(charactersOnPage, store.searchCharacter)
-            .then(data => {
-                const charactersArray: ICharacter[]  = data.data.results.map(character => {
-                    return {
-                        id: character.id,
-                        name: character.name,
-                        desc: character.description,
-                        image: character.thumbnail.path + "." + character.thumbnail.extension
-                    }
-                })
-                store.setCharacters(charactersArray);
-            })
-    }
-
     return(
         <section className={styles.characters}>
-            <SearchBar subject="Characters" amount={store.charactersAmount} inputHandler={inputHandler} callback={searchCharactersByName}
+            <SearchBar subject="Characters" amount={store.charactersAmount} inputHandler={inputHandler} callback={store.searchCharactersByName}
                        searchWord={store.searchCharacter} canselDebounce={canselDebounce}/>
             <hr className={styles.divider}/>
             {store.loading ? (
