+++ src/stores/CharactersStore.ts
@@ -1,5 +1,6 @@
 import {ICharacter} from "../types/ICharacter.tsx";
 import {makeAutoObservable} from "mobx";
+import charactersRequests from "../api/charactersRequests.ts";
 
 class CharactersStore {
     characters: ICharacter[] = [];
@@ -8,6 +9,7 @@ class CharactersStore {
     loading: boolean = false;
     charactersAmount: number = 0;
     pagesAmount: number = 0;
+    charactersOnPage : number = 36;
 
     constructor(){
         makeAutoObservable(this);
@@ -36,7 +38,56 @@ class CharactersStore {
     setPagesAmount = (pagesAmount: number) => {
         this.pagesAmount = pagesAmount;
     }
+
+    searchCharacters = (offset: number) => {
+        charactersRequests.getCharacters(this.charactersOnPage, offset)
+        .then(data => {
+            const charactersArray: ICharacter[]  = data.data.results.map(character => {
+                return {
+                    id: character.id,
+                    name: character.name,
+                    desc: character.description,
+                    image: character.thumbnail.path + "." + character.thumbnail.extension
+                }
+            })
+            this.setCharacters(charactersArray);
+            this.setCharactersAmount(data.data.total);
+            this.setPagesAmount(Math.ceil(this.charactersAmount / this.charactersOnPage));
+        })
+    }
+
+    searchCharactersByName = (event: any) => {
+        event.preventDefault();
+        charactersRequests.searchCharacterByName(this.charactersOnPage, this.searchCharacter)
+            .then(data => {
+                const charactersArray: ICharacter[]  = data.data.results.map(character => {
+                    return {
+                        id: character.id,
+                        name: character.name,
+                        desc: character.description,
+                        image: character.thumbnail.path + "." + character.thumbnail.extension
+                    }
+                })
+                this.setCharacters(charactersArray);
+            })
+    }
+
+    searchCharactersByNameDebounce = () => {
+        charactersRequests.searchCharacterByName(this.charactersOnPage, this.searchCharacter)
+            .then(data => {
+                const charactersArray: ICharacter[]  = data.data.results.map(character => {
+                    return {
+                        id: character.id,
+                        name: character.name,
+                        desc: character.description,
+                        image: character.thumbnail.path + "." + character.thumbnail.extension
+                    }
+                })
+                this.setCharacters(charactersArray);
+            })
+    }
 }
+
 
 export const charactersStore = new CharactersStore();
